tasks/2023-12-17/index.ts
export const GSL_DEMO_SNIPPET = `
Gift(ribbon: "gold curly", label: "Merry Christmas!") {
    Wearable(type: "socks", size: "small", color: "red").if(winterSeason: true) {
      pattern: "snowflakes"
    }

    Wearable(type: "scarf", size: "medium", color: "green") {
      pattern: "snowflakes"
    }

    Literary(type: "book", size: "15cm 22cm 2cm", title: "Christmas Stories", author: "C. Claus")
}
`;

export interface Gift {
	items: Wearable[]
}
export interface Wearable {
	type: string;
	color: string;
	size: string;
}

export function parseGSL(gslScript: string): Gift {
	return {};
}
